08-collab.md: Clarify who's is doing what.
The previous version was a bit confusing on which where the commands
enter by the owner and which by the collaborator.

I'm also adding some text at the beginning explaining which is the
goal of the lesson and a challenge at the end to repeat the same
process switching roles.