Update 04-changes.md
In the beginning instructions on creating and editing a file in the newly initiated repo, I have added a reminder that the bash command to open or edit text editors depends on the editor you choose. I link back to the more thorough discussion of text editors in The Unix Shell lesson (third episode). 

I find that students, particularly those new to command line interfaces, often get tripped up on opening, editing, and closing text editors. My edit is an attempt to remind students to double-check and clarify that they'll need to select their editor and the right command that goes with it.